Plumb getting/setting xattrs through InodeOperations and 9p gofer interfaces.
There was a very bare get/setxattr in the InodeOperations interface. Add
context.Context and size to both, as well as flags to setxattr.

There were existing 9p messages related to xattrs (xattrwalk and xattrcreate),
but we define our own, mainly for the sake of simplicity. Remove the old ones,
which were unused.

Extended attributes will be implemented in future commits.

PiperOrigin-RevId: 284858812