Update commandline and get local runs working.
TL;DR: The command line is updated with run-mock and
run-local commands, and run-local runs with everything
but "fio".

run with:
blaze run :benchmarks -- run-[mock|local] [workload]

Details
While updating the commandline for the new GCP command,
local runs were not working with blaze/bazel. This is
due to the scripts building Dockerfiles on the fly,
which doesn't work due to file permissions on blaze.

Thus, the changes update all workloads to Tarball rules
in their BUILD files so that they can be opened by the
scripts and passed to Docker.

fio is currently not working due to permissions errors on
mounts that it needs to create.

PiperOrigin-RevId: 286458110